Move initial view to index
