+++ src/scripts/index.js
@@ -1,9 +1,27 @@
 import '../scss/styles.scss';
-import { loadData } from './modules/db/storage';
-// eslint-disable-next-line import/no-cycle
-import initialView from './modules/views/initial-view';
+import projectView from './modules/views/project';
+import storage from './modules/db/storage';
+import defaultData from './modules/db/defaultData';
 
-const projects = loadData();
-initialView(projects);
+if (storage.projects.length === 0) {
+  localStorage.setItem('projects', JSON.stringify(defaultData));
+  storage.load();
+}
 
-export default projects;
+projectView().getUserInput();
+projectView().renderProjects();
+
+if (storage.projects.length > 0) {
+  const tabs = document.querySelectorAll('.project-tabs__item');
+  tabs[0].classList.add('active');
+  const container = document.querySelectorAll('.project-container');
+  container[0].classList.add('active');
+}
+
+const addPrBtns = document.querySelectorAll('.add-project-btn');
+addPrBtns.forEach(btn => {
+  btn.addEventListener('click', () => {
+    btn.textContent = btn.textContent === 'Add Project' ? 'Close' : 'Add Project';
+    btn.nextElementSibling.classList.toggle('open');
+  });
+});
